POAP vault registration (without backend for now)
