+++ src/components/[guild]/CreatePoap/components/MonetizePoap/index.ts
@@ -1,0 +1,3 @@
+import MonetizePoap from "./MonetizePoap"
+
+export default MonetizePoap
